refactor(desktop): extract migrated simfile Tauri commands into dedicated Rust module
Move the 5 Tauri simfile commands already migrated to typed api_contracts
(fetch_user_simfiles, get_next_display_id, fetch_cloud_song,
update_simfile_record, create_simfile_record) plus their supporting
GraphQL queries/mutations, native_simfile_from_graphql field-parsing
helpers, and preview-upload helpers out of api.rs into a new simfile.rs
module. api.rs keeps the shared GraphQL transport, Google Drive metadata
bridge, and every other command untouched. Widen success_data,
api_base_url_from_env, access_token_from_auth_state, graphql_document,
and upload_bytes_to_api to pub(crate) so simfile.rs can call them; no
other visibility or behavior changes. Split api_tests.rs accordingly
into tests/simfile_tests.rs for the moved coverage.
